Nevermind, float resolutions don't work, back to unsigned int (#54)
New resolutions: 256x224, 512x448, 768x672, 1024x896, 1280x1120, 1536x1344 and 2048x1792

+++ src/game.cpp
@@ -125,7 +125,7 @@ void Game::updateResolution() {
     if (window.isOpen()) {
         window.close();
     }
-    float resolutionMultiplier = Settings::getResolutionMultiplier();
+    unsigned int resolutionMultiplier = Settings::getResolutionMultiplier();
     window.create(sf::VideoMode(baseWidth * resolutionMultiplier,
                                 baseHeight * resolutionMultiplier),
                   "Super Mario Kart", WINDOW_STYLE);
